+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Unreleased
 ************
 * Adds `PUT /sbds/{sbd_id}/status/ready` and `PUT /sbds/{sbd_id}/status/draft`
+* Updates to `POST /panels/generate` to generate both SV panel and Science Category panels
+* Updated OSD to v5.2.1
+* Added `GET /prsls/osd/cycles` 
 
 13.1.1
 ************